Refactor doc context menu
